// src/app/model/orden-compra.model.ts
export class OrdenCompra {
    id:number;
    nroOrdenCompra: string;
    nroExpSIAF: string;
    fecha: Date;
    idTipoDocumento: number;
    nomTipoDocumento: string;
    nroDocumento: string;
    idProveedor: number;
    nomProveedor: string;
    monto: number;
    idEstado: number;
    nomEstado: string;
}